+++ fake_spectra/index_table.h
@@ -21,7 +21,7 @@ public:
   //Each element is a list of particles near that line,
   //and each list element is (particle index, distance from line).
   std::valarray< std::map<int, double> > get_near_particles(const float pos[], const float hh[], const long long npart);
-  double * assign_cells(const int i, const std::valarray< std::map<int, double> > nearby_array, const float pos[]);
+  float * assign_cells(const int i, const std::valarray< std::map<int, double> > nearby_array, const float pos[]);
 
   //Get the axis of a line
   inline int get_axis(const int iproc)
